Pass keyword arguments using the double-splat operator
Ruby 3.0 will be introducing a separation of positional and keyword
arguments. This is explained in the Ruby documentation:
https://www.ruby-lang.org/en/news/2019/12/12/separation-of-positional-and-keyword-arguments-in-ruby-3-0/

When a method includes keyword arguments, they are supplied to the
define_method/replace_method block via a hash as the last arguments
parameter.

The challenge is that a hash could be keyword arguments or it could be
a literal hash being passed as a parameter so it is necessary to check
the parameters list for keyword arguments